[op][test] Fix WAM test application build on 64bit architecture.
:Release Notes:
Fix WAM test application build on 64bit architecture.

:Detailed Notes:
Fix the long value assignment to an integer value.

:Testing Performed:
1. Configure webOS build for raspberrypi4-64.
2. Build image.
3. Wam build should pass without error.

:QA Notes:

:Issues Addressed:
[WRN-418] Include gtest into build.

Reviewed-on: http://gpro.lge.com/c/webosose/wam/+/316215
Tested-by: <aleksey.dotsenko@lgepartner.com>

+++ src/tests/utils_test.cc
@@ -153,7 +153,7 @@ TEST(UtilsTestSuite, strToInt_Overflow) {
 
 TEST(UtilsTestSuite, strToInt_MAX_VALUE) {
   std::string str = "+2147483647";
-  const int32_t expected = LONG_MAX;
+  const int32_t expected = INT_MAX;
   int value;
   bool result = util::StrToInt(str, value);
   EXPECT_TRUE(result);
@@ -169,7 +169,7 @@ TEST(UtilsTestSuite, strToInt_Underflow) {
 
 TEST(UtilsTestSuite, strToInt_MIN_VALUE) {
   std::string str = "-2147483648";
-  const int32_t expected = LONG_MIN;
+  const int32_t expected = INT_MIN;
   int value;
   bool result = util::StrToInt(str, value);
   EXPECT_TRUE(result);
